extract svg module
