+++ public/js/colors.js
@@ -1,0 +1,13 @@
+module.exports = [
+	'#807c59',
+	'#618059',
+	'#73bf60',
+	'#ccae8f',
+	'#806c59',
+	'#bf9060',
+	'#65567a',
+	'#8860bf',
+	'#bfb660',
+	'#cc8f8f',
+	'#bf6060'
+];
